requirements.txt
flask~=3.0.2
flask-mysql-connector
mysql-connector-python~=9.0.0